Corrección de espacios
