+++ src/controllers/ExtintorController.php
@@ -6,14 +6,15 @@ use Flight;
 
 use Almacen\Ssst\dbrepo\Factory;
 use Almacen\Ssst\dbrepo\RepoMain;
-use Almacen\Ssst\utils\ValExtintoress;
+
+use Almacen\Ssst\utils\ValidacionExtintores;
 use Almacen\Ssst\dbrepo\models\MExtintores;
 use DateTime;
 use Exception;
 
 class ExtintorController extends Flight{
-
-    private ValExtintoress $mextin;
+  
+    private ValidacionExtintores $extin;
     private MExtintores $model;
     private  $repoRiesgos;
     public function __construct()
@@ -21,7 +22,8 @@ class ExtintorController extends Flight{
         $this->model=new MExtintores();
         $this->repoRiesgos = new Factory();
         $this->repoRiesgos->getCatalogos(new RepoMain());
-        $this->mextin= new ValExtintoress();
+       // $this->prueba = new ValidacionExtintores();
+    $this->extin= new ValidacionExtintores();
     }
 
     public function index (){
